requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
psycopg2-binary
pydantic-settings
pydantic[email]
python-jose
python-multipart
passlib[bcrypt]
